Update beacon injection strategy to maintain logger; remove unused dev dependency; relocate DOM utils into utils/dom/
