+++ packages/browser/test/unit/specs/services/createBrowserNetworkService.spec.js
@@ -24,13 +24,13 @@ afterEach(() => {
 });
 
 describe("BrowserNetworkService", () => {
-  it("exposes sendFetchRequest and sendBeaconRequest as functions", () => {
+  it("exposes sendFetchRequest as a function and sendBeacon when available", () => {
     const network = createBrowserNetworkService();
     expect(typeof network.sendFetchRequest).toBe("function");
-    expect(typeof network.sendBeaconRequest).toBe("function");
+    expect(typeof network.sendBeacon).toBe("function");
   });
 
-  it("falls back to sendFetchRequest when navigator.sendBeacon is unavailable", () => {
+  it("exposes sendBeacon as null when navigator.sendBeacon is unavailable", () => {
     const descriptor = Object.getOwnPropertyDescriptor(
       Navigator.prototype,
       "sendBeacon",
@@ -38,7 +38,7 @@ describe("BrowserNetworkService", () => {
     delete Navigator.prototype.sendBeacon;
     try {
       const network = createBrowserNetworkService();
-      expect(network.sendBeaconRequest).toBe(network.sendFetchRequest);
+      expect(network.sendBeacon).toBeNull();
     } finally {
       if (descriptor) {
         Object.defineProperty(Navigator.prototype, "sendBeacon", descriptor);
