+++ packages/core/package.json
@@ -25,8 +25,5 @@
     "@adobe/aep-rules-engine": "^3.1.1",
     "@adobe/reactor-query-string": "^2.0.0",
     "uuid": "^14.0.0"
-  },
-  "devDependencies": {
-    "js-cookie": "3.0.7"
   }
 }
