+++ packages/core/test/unit/specs/utils/injectFireReferrerHideableImage.spec.js
@@ -11,7 +11,7 @@ governing permissions and limitations under the License.
 */
 
 import { vi, beforeEach, describe, it, expect } from "vitest";
-import injectFireReferrerHideableImage from "../../../../src/utils/injectFireReferrerHideableImage.js";
+import injectFireReferrerHideableImage from "../../../../src/utils/dom/injectFireReferrerHideableImage.js";
 
 describe("injectFireReferrerHideableImage", () => {
   let appendNodeMock;
